Work on ir_opcodes. Adding a ir_string_table and a ir_string to keep from having to manually manage memory in ir_opcode nodes.
Might need to add a "data" table instead of just a "string" table. IE. the ability to store random bytes, which i mean could technically be stored in a string as is. Maybe we don't. just add some overloaded functions or something for ease of use, ie getting a string, or a given type.
eg: i32, i64, r32, r64, etc.